create a metrics folder abd test for monai
