+++ clinicadl/utils/task_manager/classification.py
@@ -9,8 +9,8 @@ from torch.utils.data import sampler
 from torch.utils.data.distributed import DistributedSampler
 
 from clinicadl.caps_dataset.data import CapsDataset
+from clinicadl.metrics.metric_module import MetricResult
 from clinicadl.utils.exceptions import ClinicaDLArgumentError
-from clinicadl.utils.metric_module import MetricResult
 from clinicadl.utils.task_manager.task_manager import TaskManager
 
 logger = getLogger("clinicadl.task_manager")
@@ -53,7 +53,7 @@ class ClassificationManager(TaskManager):
             "MK",
             "LR_plus",
             "LR_minus",
-            "roc_auc",
+            # "roc_auc", TODO: test for monai
         ]
 
     @property
