+++ clinicadl/utils/task_manager/reconstruction.py
@@ -5,8 +5,8 @@ from torch.utils.data import sampler
 from torch.utils.data.distributed import DistributedSampler
 
 from clinicadl.caps_dataset.data import CapsDataset
+from clinicadl.metrics.metric_module import MetricResult
 from clinicadl.utils.exceptions import ClinicaDLArgumentError
-from clinicadl.utils.metric_module import MetricResult
 from clinicadl.utils.task_manager.task_manager import TaskManager
